BD is partially extracted

+++ java/refactoring/src/main/java/ru/akirakozov/sd/refactoring/DbAccessor.java
@@ -28,7 +28,7 @@ public class DbAccessor {
             " PRICE          INT     NOT NULL)";
 
     @Contract(pure = true)
-    public static @NotNull String insert(String name, int price) {
+    public static @NotNull String insert(String name, long price) {
         return "INSERT INTO PRODUCT (NAME, PRICE) VALUES (\"" + name + "\", " + price + ")";
     }
